🎨 extract Source in module
