+++ src/module/mod.rs
@@ -25,7 +25,8 @@ mod test_deserialize {
     use crate::module::gen_mod::{GeneratedMod, GenModComponent};
     use crate::module::location::github::{Github, GithubDescriptor};
     use crate::module::location::http::Http;
-    use crate::module::location::{ConcreteLocation, Source};
+    use crate::module::location::ConcreteLocation;
+    use crate::module::location::source::Source;
     use crate::module::module_conf::{ModuleConf, ModuleContent};
     use crate::module::weidu_mod::WeiduMod;
     use crate::post_install::PostInstall;
